Update README.md for dotfiles 4
Seeing as how I've had a SSD error and needed to buy a new one, this is missing the proper `.bash_aliases` file that should be the corner stone of this entire dotfile. I have place a disclaimer warning and explaining the situation and will remake a proper and correct file to go along with this dotfile. I have remade the aliases to match the provided examples and fixed the documentation to be more accurate to the information used and needed cited.